[Plugin] enh: reformat code

+++ openmole/plugins/org.openmole.plugin.task.r/src/main/scala/org/openmole/plugin/task/r/RTask.scala
@@ -74,14 +74,13 @@ object RTask {
     def version = rVersion
 
     // add additional installation of devtools only if needed
-    val installCommands = if (libraries.exists { case l: InstallCommand.RLibrary ⇒ l.version.isDefined }) {
-      install ++ Seq("apt update", "apt-get -y install libssl-dev libxml2-dev libcurl4-openssl-dev libssh2-1-dev",
-        """R --slave -e 'install.packages("devtools", dependencies = T); library(devtools);""") ++
-        InstallCommand.installCommands(libraries.toVector ++ Seq(InstallCommand.RLibrary("jsonlite", None)))
-    }
-    else {
-      install ++ InstallCommand.installCommands(libraries.toVector ++ Seq(InstallCommand.RLibrary("jsonlite", None)))
-    }
+    val installCommands =
+      if (libraries.exists { case l: InstallCommand.RLibrary ⇒ l.version.isDefined }) {
+        install ++ Seq("apt update", "apt-get -y install libssl-dev libxml2-dev libcurl4-openssl-dev libssh2-1-dev",
+          """R --slave -e 'install.packages("devtools", dependencies = T); library(devtools);""") ++
+          InstallCommand.installCommands(libraries.toVector ++ Seq(InstallCommand.RLibrary("jsonlite", None)))
+      }
+      else install ++ InstallCommand.installCommands(libraries.toVector ++ Seq(InstallCommand.RLibrary("jsonlite", None)))
 
     val uDockerArguments =
       UDockerTask.createUDocker(
